Move `/me/` handlers to the separate file
